Added lemmatisation
The previous code searched for synsets for an input word like
'pictures' or 'pulverised', which have suffixes like 's' and 'ed',

The synset search of Wordnet automatically lemmatises the input to
find matches like 'picture' and 'pulverise' (the altered forms not
being stored in the dictionary), so these searches will produce
results, but the code checking that the returned synset has the
intended content doesn't allow such matches.

Adding in 'wn.morphy(word)' as an allowed match makes the matching
more permissive and lets long plurals and past tenses be transformed.

This results in better compression for the input text, though obviously
also adds to the 'loss' because it permits shifts in tense.